Update README.md
This is a content-based movie recommendation system built using machine learning. It is capable of suggesting the movie according to user preferences. You can Search and Filter the movie according to your needs. The most unique and fresh feature of this project is that it used the IP Address api which will suggest the movie based on the User's location and region.